refactor(eval): extract judge.rs with batch API support
Moves LLM-as-judge types (JudgmentTuple, JudgmentResult,
JudgedApproachResult, JudgedE2EReport) and functions (judge_with_claude,
judge_with_claude_model, judge_single_tuple, judge_single_tuple_model,
parse_judge_json, aggregate_judgments, save/load_judgment_tuples) from
token_efficiency.rs into a new judge.rs module.

Adds task-specific prompt functions ported from longmemeval.rs and
locomo.rs: lme_anscheck_prompt, lme_answer_prompt, locomo_judge_prompt.

Adds judge_with_batch_api — a new function using the Anthropic Batch API
(50% cheaper, no rate limits) to judge JudgmentTuples at scale.

Removes the feature-gated eval_judge stub from mod.rs (it was empty).
token_efficiency.rs re-exports everything for backward compat.
