+++ crates/origin-core/src/eval/token_efficiency.rs
@@ -2478,304 +2478,12 @@ pub struct E2ELocomoReport {
     pub results: Vec<E2ELocomoResult>,
 }
 
-// ===== LLM-as-Judge Types =====
-
-/// A single E2E answer to be judged.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct JudgmentTuple {
-    pub question: String,
-    pub ground_truth: String,
-    pub approach: String,
-    pub answer: String,
-    pub context_tokens: usize,
-}
-
-/// Result from the LLM judge.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct JudgmentResult {
-    pub question: String,
-    pub approach: String,
-    /// 0 or 1.
-    pub score: u8,
-    pub reason: String,
-    pub context_tokens: usize,
-}
-
-/// Per-approach aggregated result in a judged E2E report.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct JudgedApproachResult {
-    pub approach: String,
-    /// Fraction of questions scoring 1.
-    pub accuracy: f64,
-    pub total: usize,
-    pub correct: usize,
-    pub mean_context_tokens: f64,
-}
-
-/// Full judged E2E report.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct JudgedE2EReport {
-    pub judge_model: String,
-    pub total_judged: usize,
-    pub results_by_approach: Vec<JudgedApproachResult>,
-}
-
-// ===== LLM-as-Judge Functions =====
-
-/// Save E2E answer tuples to JSON for offline judging.
-pub fn save_judgment_tuples(tuples: &[JudgmentTuple], path: &Path) -> Result<(), std::io::Error> {
-    let json = serde_json::to_string_pretty(tuples).map_err(std::io::Error::other)?;
-    std::fs::write(path, json)
-}
-
-/// Load previously saved judgment tuples from JSON.
-pub fn load_judgment_tuples(path: &Path) -> Result<Vec<JudgmentTuple>, std::io::Error> {
-    let content = std::fs::read_to_string(path)?;
-    serde_json::from_str(&content).map_err(std::io::Error::other)
-}
-
-/// Judge answer tuples using Claude via the `claude -p` CLI.
-///
-/// Requires Claude Code CLI installed (`claude --version` must succeed).
-/// Uses Haiku via the user's existing Max subscription — no API key needed.
-/// Runs up to `concurrency` judgments in parallel.
-pub async fn judge_with_claude(
-    tuples: &[JudgmentTuple],
-    concurrency: usize,
-) -> Result<Vec<JudgmentResult>, OriginError> {
-    judge_with_claude_model(tuples, concurrency, "haiku").await
-}
-
-/// Judge tuples with a specific Claude model (e.g. "haiku", "sonnet").
-pub async fn judge_with_claude_model(
-    tuples: &[JudgmentTuple],
-    concurrency: usize,
-    model: &str,
-) -> Result<Vec<JudgmentResult>, OriginError> {
-    use tokio::sync::Semaphore;
-
-    let semaphore = Arc::new(Semaphore::new(concurrency));
-    let mut handles = Vec::new();
-
-    for tuple in tuples {
-        let sem = semaphore.clone();
-        let tuple = tuple.clone();
-        let model = model.to_string();
-
-        let handle = tokio::spawn(async move {
-            let _permit = sem.acquire().await.unwrap();
-            judge_single_tuple_model(&tuple, &model).await
-        });
-        handles.push(handle);
-    }
-
-    let mut results = Vec::new();
-    for handle in handles {
-        match handle.await {
-            Ok(Ok(result)) => results.push(result),
-            Ok(Err(e)) => log::warn!("[judge] judgment failed: {}", e),
-            Err(e) => log::warn!("[judge] task panicked: {}", e),
-        }
-    }
-
-    Ok(results)
-}
-
-/// Judge a single (question, ground_truth, answer) tuple via `claude -p`.
-///
-/// Passes the prompt via stdin and disables all tools (`--allowedTools ""`), which
-/// prevents Claude Code's agentic tool-calling loop and gets a direct text/JSON response.
-/// OAuth auth from the user's existing login is used (no API key required).
-pub async fn judge_single_tuple(tuple: &JudgmentTuple) -> Result<JudgmentResult, OriginError> {
-    judge_single_tuple_model(tuple, "haiku").await
-}
-
-/// Judge a single tuple with a specific Claude model.
-pub async fn judge_single_tuple_model(
-    tuple: &JudgmentTuple,
-    model: &str,
-) -> Result<JudgmentResult, OriginError> {
-    use tokio::io::AsyncWriteExt;
-    use tokio::process::Command;
-
-    let prompt = format!(
-        "Question: {}\n\nReference answer: {}\n\nCandidate answer: {}\n\nIs the candidate answer correct? Compare against the reference.",
-        tuple.question, tuple.ground_truth, tuple.answer
-    );
-
-    let json_schema = r#"{"type":"object","properties":{"score":{"type":"integer","enum":[0,1]},"reason":{"type":"string"}},"required":["score","reason"]}"#;
-
-    let system_prompt = "You are an expert evaluator judging answer correctness. \
-        Score 1 if the candidate answer contains the key information from the reference answer \
-        (even if worded differently). Score 0 if the candidate answer is wrong, missing key \
-        information, or irrelevant. Think step by step before scoring.";
-
-    let mut child = Command::new("claude")
-        .args([
-            "-p",
-            "--model",
-            model,
-            "--output-format",
-            "json",
-            "--json-schema",
-            json_schema,
-            "--system-prompt",
-            system_prompt,
-            "--no-session-persistence",
-            "--allowedTools",
-            "",
-        ])
-        .stdin(std::process::Stdio::piped())
-        .stdout(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::piped())
-        .spawn()
-        .map_err(|e| OriginError::Generic(format!("claude -p failed to start: {}", e)))?;
-
-    // Write prompt to stdin then close it.
-    if let Some(mut stdin) = child.stdin.take() {
-        stdin
-            .write_all(prompt.as_bytes())
-            .await
-            .map_err(|e| OriginError::Generic(format!("write to claude stdin failed: {}", e)))?;
-        // drop closes stdin
-    }
-
-    let output = child
-        .wait_with_output()
-        .await
-        .map_err(|e| OriginError::Generic(format!("claude -p wait failed: {}", e)))?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        let stdout_preview = String::from_utf8_lossy(&output.stdout);
-        return Err(OriginError::Generic(format!(
-            "claude -p exited with error: stderr={} stdout={}",
-            stderr.trim(),
-            stdout_preview.trim()
-        )));
-    }
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-
-    // Parse the JSON response. `--output-format json` returns an envelope; the structured
-    // output lives in the `structured_output` field when `--json-schema` is used.
-    let parsed: serde_json::Value = parse_judge_json(&stdout).map_err(|e| {
-        OriginError::Generic(format!(
-            "judge response parse error: {} — raw: {}",
-            e, stdout
-        ))
-    })?;
-
-    let score = parsed.get("score").and_then(|v| v.as_u64()).unwrap_or(0) as u8;
-    let reason = parsed
-        .get("reason")
-        .and_then(|v| v.as_str())
-        .unwrap_or("no reason")
-        .to_string();
-
-    Ok(JudgmentResult {
-        question: tuple.question.clone(),
-        approach: tuple.approach.clone(),
-        score,
-        reason,
-        context_tokens: tuple.context_tokens,
-    })
-}
-
-/// Try to extract the judgment JSON object from `claude -p` output.
-///
-/// When `--output-format json` is combined with `--json-schema`, Claude Code returns an
-/// envelope like:
-/// ```json
-/// {"type":"result", "structured_output": {"score":1, "reason":"..."}, ...}
-/// ```
-/// We try several strategies to locate the score/reason object:
-/// 1. `structured_output` field in the envelope (primary path).
-/// 2. `result` field in the envelope (fallback for older CLI versions).
-/// 3. Top-level object if it already contains `score`.
-/// 4. Extract any `{...}` substring (last resort).
-fn parse_judge_json(stdout: &str) -> Result<serde_json::Value, serde_json::Error> {
-    let trimmed = stdout.trim();
-
-    if let Ok(envelope) = serde_json::from_str::<serde_json::Value>(trimmed) {
-        // Strategy 1: structured_output field (primary — used when --json-schema is set).
-        if let Some(so) = envelope.get("structured_output") {
-            if so.get("score").is_some() {
-                return Ok(so.clone());
-            }
-        }
-        // Strategy 2: result field (text-mode fallback).
-        if let Some(result) = envelope.get("result") {
-            if result.get("score").is_some() {
-                return Ok(result.clone());
-            }
-            // result may be a JSON string — try to parse it.
-            if let Some(s) = result.as_str() {
-                if let Ok(inner) = serde_json::from_str::<serde_json::Value>(s) {
-                    if inner.get("score").is_some() {
-                        return Ok(inner);
-                    }
-                }
-            }
-        }
-        // Strategy 3: top-level already has score.
-        if envelope.get("score").is_some() {
-            return Ok(envelope);
-        }
-    }
-
-    // Strategy 4: extract the first balanced {...} block.
-    if let (Some(start), Some(end)) = (trimmed.find('{'), trimmed.rfind('}')) {
-        if start <= end {
-            if let Ok(v) = serde_json::from_str::<serde_json::Value>(&trimmed[start..=end]) {
-                if v.get("score").is_some() {
-                    return Ok(v);
-                }
-            }
-        }
-    }
-
-    // Final fallback: return a parse error to surface the raw output.
-    serde_json::from_str(trimmed)
-}
-
-/// Aggregate judgment results into a report sorted by accuracy descending.
-pub fn aggregate_judgments(results: &[JudgmentResult], judge_model: &str) -> JudgedE2EReport {
-    let mut by_approach: HashMap<String, Vec<&JudgmentResult>> = HashMap::new();
-    for r in results {
-        by_approach.entry(r.approach.clone()).or_default().push(r);
-    }
-
-    let mut approach_results: Vec<JudgedApproachResult> = by_approach
-        .iter()
-        .map(|(approach, items)| {
-            let total = items.len();
-            let correct = items.iter().filter(|r| r.score == 1).count();
-            let accuracy = correct as f64 / total.max(1) as f64;
-            let mean_tokens =
-                items.iter().map(|r| r.context_tokens as f64).sum::<f64>() / total.max(1) as f64;
-            JudgedApproachResult {
-                approach: approach.clone(),
-                accuracy,
-                total,
-                correct,
-                mean_context_tokens: mean_tokens,
-            }
-        })
-        .collect();
-
-    approach_results.sort_by(|a, b| {
-        b.accuracy
-            .partial_cmp(&a.accuracy)
-            .unwrap_or(std::cmp::Ordering::Equal)
-    });
-
-    JudgedE2EReport {
-        judge_model: judge_model.to_string(),
-        total_judged: results.len(),
-        results_by_approach: approach_results,
-    }
-}
+// Re-exports from judge.rs (backward compat)
+pub use crate::eval::judge::{
+    aggregate_judgments, judge_single_tuple, judge_single_tuple_model, judge_with_claude,
+    judge_with_claude_model, load_judgment_tuples, parse_judge_json, save_judgment_tuples,
+    JudgedApproachResult, JudgedE2EReport, JudgmentResult, JudgmentTuple,
+};
 
 /// Score an LLM answer against a ground-truth string using keyword overlap.
 ///
